merge @eauge implementation into new solution file
the old solution file is still present but the 'Sdk' project has been
removed since it is completely transferred into the new solution.

the binaries in the packages directory have been removed

there is a possible issue at this commit with package restore. an older
MSBuild-Integrated Package Restore option was breifly enabled on the
solution. afterwards it was removed but there may still be remnants of
enabling this. the preferred solution is now Automatic Package Restore
in Visual Studio. See http://docs.nuget.org/docs/reference/package-restore
for more details.