renames namespace to OpenTokSDK. adds OpenTokTest project
Apparently, for many reasons it is not okay to name an object inside
a namespace the same as the enclosing namespace. In light of this,
renaming the namespace rather than the main class seems like it is less
disruptive since it should only effect 'using' statements. Most places
in a developer's code will still be able to use the shorter class
name OpenTok.